+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_form_training_client_async.py
@@ -81,15 +81,16 @@ class FormTrainingClient(FormRecognizerClientBaseAsync):
     ) -> None:
         api_version = kwargs.pop("api_version", FormRecognizerApiVersion.V2_1)
         super(FormTrainingClient, self).__init__(
-            endpoint=endpoint, credential=credential, api_version=api_version, client_kind="form", **kwargs
+            endpoint=endpoint,
+            credential=credential,
+            api_version=api_version,
+            client_kind="form",
+            **kwargs
         )
 
     @distributed_trace_async
     async def begin_training(
-        self,
-        training_files_url: str,
-        use_training_labels: bool,
-        **kwargs: Any
+        self, training_files_url: str, use_training_labels: bool, **kwargs: Any
     ) -> AsyncLROPoller[CustomFormModel]:
         """Create and train a custom model. The request must include a `training_files_url` parameter that is an
         externally accessible Azure storage blob container URI (preferably a Shared Access Signature URI). Note that
@@ -202,7 +203,9 @@ class FormTrainingClient(FormRecognizerClientBaseAsync):
             cls=deserialization_callback,
             continuation_token=continuation_token,
             polling=AsyncLROBasePolling(
-                timeout=polling_interval, lro_algorithms=[FormTrainingPolling()], **kwargs
+                timeout=polling_interval,
+                lro_algorithms=[FormTrainingPolling()],
+                **kwargs
             ),
             **kwargs
         )