+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added `status` attribute to `LogsQueryResult`.
 - Added `LogsQueryStatus` Enum to describe the status of a result.
 - Added a new `LogsTableRow` type that represents a single row in a table.
-- Items in `metrics` list in `MetricsResult` can now be accessed by metric names.
+- Items in `metrics` list in `MetricsQueryResult` can now be accessed by metric names.
 - Added `audience` keyword to support providing credential scope when creating clients.
 
 ### Breaking Changes
@@ -19,6 +19,7 @@
 - `query_workspace` API now returns a union of `LogsQueryPartialResult` and `LogsQueryResult`.
 - `query_batch` API now returns a union of `LogsQueryPartialResult`, `LogsQueryError` and `LogsQueryResult`.
 - `metric_namespace` is renamed to `namespace` and is a keyword-only argument in `list_metric_definitions` API.
+- `MetricsResult` is renamed to `MetricsQueryResult`.
 
 ## 1.0.0b4 (2021-09-09)